+++ src/MainWindow.cpp
@@ -83,7 +83,7 @@ void MainWindow::initWidgets()
 void MainWindow::stylingMenu()
 {
 	// https://doc.qt.io/qt-5/stylesheet-examples.html
-	ui->toolBar->setStyleSheet( "QToolBar::separator { width: 1px; background: black; margin: 2px 10px 2px 10px; }" );
+	ui->toolBar->setStyleSheet( "QToolBar::separator { width: 1px; background: black; margin: 3px 7px 3px 7px; }" );
 	ui->menuBar->setStyleSheet( "QMenu::separator { height: 1px; background: black; margin: 7px 3px 7px 3px; }" );
 }